+++ packages/subscription-controller/src/SubscriptionController.ts
@@ -33,10 +33,12 @@ import type {
   CachedLastSelectedPaymentMethod,
   SubmitSponsorshipIntentsMethodParams,
   RecurringInterval,
+  SubscriptionStatus,
 } from './types';
 import {
   PAYMENT_TYPES,
   PRODUCT_TYPES,
+  SUBSCRIPTION_STATUSES,
   type ISubscriptionService,
   type PricingResponse,
   type ProductType,
@@ -509,26 +511,50 @@ export class SubscriptionController extends StaticIntervalPollingController()<
       lastSelectedPaymentMethod[PRODUCT_TYPES.SHIELD];
     this.#assertIsPaymentMethodCrypto(lastSelectedPaymentMethodShield);
 
-    const isTrialed = trialedProducts?.includes(PRODUCT_TYPES.SHIELD);
-
     const productPrice = this.#getProductPriceByProductAndPlan(
       PRODUCT_TYPES.SHIELD,
       lastSelectedPaymentMethodShield.plan,
     );
+    const isTrialed = trialedProducts?.includes(PRODUCT_TYPES.SHIELD);
+    // get the latest subscriptions state to check if the user has an active shield subscription
+    await this.getSubscriptions();
+    const currentSubscription = this.state.subscriptions.find((subscription) =>
+      subscription.products.some((p) => p.name === PRODUCT_TYPES.SHIELD),
+    );
 
-    const params = {
-      products: [PRODUCT_TYPES.SHIELD],
-      isTrialRequested: !isTrialed,
-      recurringInterval: productPrice.interval,
-      billingCycles: productPrice.minBillingCycles,
-      chainId,
-      payerAddress: txMeta.txParams.from as Hex,
-      tokenSymbol: lastSelectedPaymentMethodShield.paymentTokenSymbol,
-      rawTransaction: rawTx as Hex,
-      isSponsored,
-      useTestClock: lastSelectedPaymentMethodShield.useTestClock,
-    };
-    await this.startSubscriptionWithCrypto(params);
+    this.#assertValidSubscriptionStateForCryptoApproval({
+      product: PRODUCT_TYPES.SHIELD,
+    });
+    // if shield subscription exists, this transaction is for changing payment method
+    const isChangePaymentMethod = Boolean(currentSubscription);
+
+    if (isChangePaymentMethod) {
+      await this.updatePaymentMethod({
+        paymentType: PAYMENT_TYPES.byCrypto,
+        subscriptionId: (currentSubscription as Subscription).id,
+        chainId,
+        payerAddress: txMeta.txParams.from as Hex,
+        tokenSymbol: lastSelectedPaymentMethodShield.paymentTokenSymbol,
+        rawTransaction: rawTx as Hex,
+        recurringInterval: productPrice.interval,
+        billingCycles: productPrice.minBillingCycles,
+      });
+    } else {
+      const params = {
+        products: [PRODUCT_TYPES.SHIELD],
+        isTrialRequested: !isTrialed,
+        recurringInterval: productPrice.interval,
+        billingCycles: productPrice.minBillingCycles,
+        chainId,
+        payerAddress: txMeta.txParams.from as Hex,
+        tokenSymbol: lastSelectedPaymentMethodShield.paymentTokenSymbol,
+        rawTransaction: rawTx as Hex,
+        isSponsored,
+        useTestClock: lastSelectedPaymentMethodShield.useTestClock,
+      };
+      await this.startSubscriptionWithCrypto(params);
+    }
+
     // update the subscriptions state after subscription created in server
     await this.getSubscriptions();
   }
@@ -797,6 +823,34 @@ export class SubscriptionController extends StaticIntervalPollingController()<
       throw new Error(SubscriptionControllerErrorMessage.ProductPriceNotFound);
     }
     return productPrice;
+  }
+
+  #assertValidSubscriptionStateForCryptoApproval({
+    product,
+  }: {
+    product: ProductType;
+  }) {
+    const subscription = this.state.subscriptions.find((sub) =>
+      sub.products.some((p) => p.name === product),
+    );
+
+    const isValid =
+      !subscription ||
+      (
+        [
+          SUBSCRIPTION_STATUSES.pastDue,
+          SUBSCRIPTION_STATUSES.unpaid,
+          SUBSCRIPTION_STATUSES.paused,
+          SUBSCRIPTION_STATUSES.provisional,
+          SUBSCRIPTION_STATUSES.active,
+          SUBSCRIPTION_STATUSES.trialing,
+        ] as SubscriptionStatus[]
+      ).includes(subscription.status);
+    if (!isValid) {
+      throw new Error(
+        SubscriptionControllerErrorMessage.SubscriptionNotValidForCryptoApproval,
+      );
+    }
   }
 
   #assertIsUserNotSubscribed({ products }: { products: ProductType[] }) {