+++ packages/transaction-pay-controller/src/strategy/relay/relay-quotes.test.ts
@@ -6,7 +6,6 @@ import type {
 import type { Hex } from '@metamask/utils';
 import { cloneDeep } from 'lodash';
 
-import { RELAY_URL_QUOTE } from './constants';
 import { getRelayQuotes } from './relay-quotes';
 import type { RelayQuote } from './types';
 import {
@@ -20,6 +19,7 @@ import type {
   GetDelegationTransactionCallback,
   QuoteRequest,
 } from '../../types';
+import { DEFAULT_RELAY_QUOTE_URL } from '../../utils/feature-flags';
 import {
   calculateGasCost,
   calculateGasFeeTokenCost,
@@ -205,7 +205,7 @@ describe('Relay Quotes Utils', () => {
       });
 
       expect(successfulFetchMock).toHaveBeenCalledWith(
-        RELAY_URL_QUOTE,
+        DEFAULT_RELAY_QUOTE_URL,
         expect.objectContaining({
           body: JSON.stringify({
             amount: QUOTE_REQUEST_MOCK.targetAmountMinimum,
@@ -644,6 +644,46 @@ describe('Relay Quotes Utils', () => {
             value: '0x0',
           }),
         );
+      });
+
+      it('not using gas fee token cost if chain disabled in feature flag', async () => {
+        successfulFetchMock.mockResolvedValue({
+          json: async () => QUOTE_MOCK,
+        } as never);
+
+        getTokenBalanceMock.mockReturnValue('1724999999999999');
+        getGasFeeTokensMock.mockResolvedValue([GAS_FEE_TOKEN_MOCK]);
+
+        getRemoteFeatureFlagControllerStateMock.mockReturnValue({
+          cacheTimestamp: 0,
+          remoteFeatureFlags: {
+            confirmations_pay: {
+              relayDisabledGasStationChains: [QUOTE_REQUEST_MOCK.sourceChainId],
+            },
+          },
+        });
+
+        const result = await getRelayQuotes({
+          messenger,
+          requests: [QUOTE_REQUEST_MOCK],
+          transaction: TRANSACTION_META_MOCK,
+        });
+
+        expect(result[0].fees.isSourceGasFeeToken).toBeUndefined();
+        expect(result[0].fees.sourceNetwork).toStrictEqual({
+          estimate: {
+            fiat: '4.56',
+            human: '1.725',
+            raw: '1725000000000000',
+            usd: '3.45',
+          },
+          max: {
+            fiat: '4.56',
+            human: '1.725',
+            raw: '1725000000000000',
+            usd: '3.45',
+          },
+        });
       });
     });